mpylib: Bytecode: When processing qstr-taking bytecode ops, store actual qstr.
Original purpose of Bytecode class was generation content for .mpy writing,
so zeroes were stored as argument of bytecode op, and actual qstr was stored
in .co_names array. But another usage is to generate in-memory bytecode, so
actual qstr value should be stored in the instruction.

However, add a global switch to keep old behavior. (It's actually less useful
with .mpy v4, where zero qstr bytes won't be written into .mpy, and instead
actual qstr string value will be. But with this switch on, at least sys.intern
won't be called, which may be helpful for some configs.)